third_party/blink/web_tests/external/wpt/css/css-content/parsing/content-valid-expected.txt

This is a testharness.js-based test.
Found 8 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] e.style['content'] = "counter(counter-name, dECiMaL) / \\"alt text\\" attr(foo) \\"bar\\"" should set the property value
  assert_equals: serialization should be canonical expected "counter(counter-name) / \\"alt text\\" attr(foo) \\"bar\\"" but got "counter(counter-name, dECiMaL) / \\"alt text\\" attr(foo) \\"bar\\""
[FAIL] e.style['content'] = "counter(counter-name, dECiMaL) / attr(foo)" should set the property value
  assert_equals: serialization should be canonical expected "counter(counter-name) / attr(foo)" but got "counter(counter-name, dECiMaL) / attr(foo)"
[FAIL] e.style['content'] = "counter(counter-name, DECIMAL) / \\"alt text\\" attr(foo) \\"bar\\"" should set the property value
  assert_equals: serialization should be canonical expected "counter(counter-name) / \\"alt text\\" attr(foo) \\"bar\\"" but got "counter(counter-name, DECIMAL) / \\"alt text\\" attr(foo) \\"bar\\""
[FAIL] e.style['content'] = "counter(counter-name, DECIMAL) / attr(foo)" should set the property value
  assert_equals: serialization should be canonical expected "counter(counter-name) / attr(foo)" but got "counter(counter-name, DECIMAL) / attr(foo)"
[FAIL] e.style['content'] = "counters(counter-name, \\".\\", dECiMaL) / \\"alt text\\" attr(foo) \\"bar\\"" should set the property value
  assert_equals: serialization should be canonical expected "counters(counter-name, \\".\\") / \\"alt text\\" attr(foo) \\"bar\\"" but got "counters(counter-name, \\".\\", dECiMaL) / \\"alt text\\" attr(foo) \\"bar\\""
[FAIL] e.style['content'] = "counters(counter-name, \\".\\", dECiMaL) / attr(foo)" should set the property value
  assert_equals: serialization should be canonical expected "counters(counter-name, \\".\\") / attr(foo)" but got "counters(counter-name, \\".\\", dECiMaL) / attr(foo)"
[FAIL] e.style['content'] = "counters(counter-name, \\".\\", DECIMAL) / \\"alt text\\" attr(foo) \\"bar\\"" should set the property value
  assert_equals: serialization should be canonical expected "counters(counter-name, \\".\\") / \\"alt text\\" attr(foo) \\"bar\\"" but got "counters(counter-name, \\".\\", DECIMAL) / \\"alt text\\" attr(foo) \\"bar\\""
[FAIL] e.style['content'] = "counters(counter-name, \\".\\", DECIMAL) / attr(foo)" should set the property value
  assert_equals: serialization should be canonical expected "counters(counter-name, \\".\\") / attr(foo)" but got "counters(counter-name, \\".\\", DECIMAL) / attr(foo)"
Harness: the test ran to completion.
